Fixed build error on amd64 (due to qMin mixing ints and longs)
svn path=/trunk/extragear/multimedia/k3b/; revision=741862

+++ libk3b/projects/k3bpipebuffer.cpp
@@ -212,7 +212,7 @@ public:
   }
 
   char* buffer;
-  size_t bufSize;
+  unsigned int bufSize;
   int outFd;
   int inFd;
   int inFdPair[2];
